Relocated files
-Relocated files to match home-page branch folder structure
-Added additional path to tsconfig for importing assets from new file locations
