--- src/components/Mission_Statement/index.tsx
@@ -1,1 +1,0 @@
-export {default as Action_Block} from "./Action_Block"
